Move static_fused_moe to vllm.hpu.ops
