+++ vllm/hpu/ops.py
@@ -12,6 +12,7 @@ import habana_frameworks.torch as htorch
 from typing import List, Optional, Tuple
 
 import vllm.hpu.utils as hpu_utils
+from vllm.model_executor.layers.activation import SiluAndMul
 
 PA_SPLIT_VALUE = (os.environ.get('PA_SPLIT_VALUE', '0') == '1')
 
@@ -113,3 +114,31 @@ def apply_rope(
 
 def awq_gemm(*args):
     raise NotImplementedError
+
+
+@hpu_utils.with_mark_steps
+def static_fused_moe(hidden_states, w1, w2, score, topk):
+    B, D = hidden_states.shape
+    num_experts = w1.shape[0]
+    routing_weights = F.softmax(score, dim=1, dtype=torch.float32)
+    routing_weights, selected_experts = torch.topk(routing_weights, topk, dim=-1)
+    routing_weights = routing_weights.to(hidden_states.dtype)
+    final_hidden_states = torch.zeros(
+            (1, B, D), dtype=hidden_states.dtype, device=hidden_states.device
+    )
+    padded_weights = torch.zeros(
+            (B, num_experts), dtype=hidden_states.dtype, device=hidden_states.device
+    )
+    padded_weights.scatter_(-1, selected_experts, routing_weights)
+    padded_weights = padded_weights.reshape(-1, B, w1.shape[0])
+    padded_weights = padded_weights.permute(2, 0, 1).unsqueeze(-1)
+
+    for expert_idx in range(num_experts):
+        padded_weight = padded_weights[expert_idx]
+        current_state_static = hidden_states.reshape(-1, D)
+        w_output = SiluAndMul()(torch.matmul(current_state_static, w1[expert_idx].transpose(0, 1)))
+        w_output = torch.matmul(w_output, w2[expert_idx].transpose(0, 1))
+        current_hidden_states_static = w_output * padded_weight
+        final_hidden_states += current_hidden_states_static
+
+    return final_hidden_states.view(-1, D)
